backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
pydantic[email]
python-dotenv
gunicorn
uvicorn[standard]
